+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Make the `filename` argument optional
+
 0.0.0 / 2014-09-18
 ==================